MVC Compelete (User = CRUD, Car = CRUD, Group = RUD)

+++ app/controllers/groups_controller.rb
@@ -13,6 +13,7 @@ class GroupsController < ApplicationController
         @group = Group.find(params[:id])
         
         if @group.update(group_params)
+            byebug
             update_user_group(@group, params[:group][:user_ids])
             flash[:success] = 'group information updated successfully!'
         else
@@ -40,16 +41,16 @@ class GroupsController < ApplicationController
     def update_user_group (group, new_user_ids)
         # byebug
         UserGroup.delete_by_group(group)
-        # new_user_ids.shift
-        @new_list_of_users = new_user_ids
+        new_user_ids.shift
         # byebug
-        @new_list_of_users.each do |user_id|
+        new_user_ids.each do |user_id|
             updated_user_group = UserGroup.new(user_id: user_id, group: group)
             updated_user_group.save
             # byebug
-            # user = User.find(user_id)
-            # byebug
-            # UserGroup.delete_by_group(user.groups)
+            puts ''
+            user = User.find(user_id)
+            byebug
+            delete_empty_groups(user.groups)
         end
     end
 
@@ -64,11 +65,13 @@ class GroupsController < ApplicationController
     end
 
     def delete_empty_groups(groups)
-            empty_group = groups.select {|group| group.name == "Not in a Group"}
-        if empty_group.length == 1
-            empty_group_join = UserGroup.find_by(group: empty_group)
             # byebug
-            empty_group.destroy
+            empty_group = groups.select {|group| group.name == "Not in a Group"}
+
+        if empty_group.length == 1 && empty_group.length != 0
+            empty_group_join = UserGroup.find_by(group: empty_group)
+            byebug
+            empty_group.first.destroy
             empty_group_join.destroy
         end
     end
